Downgrade path package to >= 1.6.4 to solve conflict with flutter_test Clean up auto_route index file
