+++ auto_route/lib/auto_route.dart
@@ -1,10 +1,5 @@
 library auto_route;
 
-import 'package:auto_route/src/extended_navigator.dart';
-import 'package:flutter/widgets.dart' show BuildContext;
-
-import 'src/route_guard.dart';
-
 export 'package:meta/meta.dart' show required;
 export 'src/auto_route_wrapper.dart';
 export 'src/extended_navigator.dart';
@@ -15,12 +10,3 @@ export 'src/route_guard.dart';
 export 'src/router_base.dart';
 export 'src/router_utils.dart';
 export 'src/transitions_builders.dart';
-
-typedef OnNavigationRejected = void Function(RouteGuard guard);
-
-extension BuildContextX on BuildContext {
-  ExtendedNavigatorState get navigator =>
-      ExtendedNavigator.of(this, nullOk: true);
-  ExtendedNavigatorState get rootNavigator =>
-      ExtendedNavigator.of(this, rootRouter: true, nullOk: true);
-}
