+++ auto_route/lib/src/router_base.dart
@@ -33,8 +33,6 @@ abstract class RouterBase {
         data = RouteData(match);
       }
       return pagesMap[match.routeDef.page](data);
-=======
-
     }
     return null;
   }
