+++ auto_route/pubspec.yaml
@@ -1,6 +1,6 @@
 name: auto_route
 description: AutoRoute is a declarative routing solution, where everything needed for navigation is automatically generated for you.
-version: 0.6.4
+version: 0.6.5
 homepage: https://github.com/Milad-Akarie/auto_route_library
 
 environment:
@@ -10,7 +10,6 @@ dependencies:
   flutter:
     sdk: flutter
   meta: ^1.1.8
-  path: ^1.7.0
 dev_dependecies:
   flutter_test:
     sdk: flutter
