+++ auto_route_generator/pubspec.yaml
@@ -1,6 +1,6 @@
 name: auto_route_generator
 description: AutoRoute is a declartive routing solution, where everything needed for navigation is automatically generated for you.
-version: 0.6.4
+version: 0.6.5
 homepage: https://github.com/Milad-Akarie/auto_route_library
 
 environment:
@@ -10,9 +10,9 @@ dependencies:
   build: ^1.0.0
   source_gen: ^0.9.0
   analyzer: ">=0.39.2 <0.40.0"
-  path: ^1.7.0
+  path: ^1.6.4
   auto_route:
-     ^0.6.4
+     ^0.6.5
 #    path: ../auto_route
 
 dev_dependencies:
