+++ example/lib/main.dart
@@ -1,8 +1,8 @@
+import 'package:auto_route/auto_route.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 
 import 'router/route_guards.dart';
-import 'router/router.dart';
 import 'router/router.gr.dart';
 
 void main() {
