+++ example/lib/router/router.dart
@@ -9,14 +9,13 @@ import '../screens/users/sub/sub/posts_details.dart';
 import '../screens/users/sub/sub/posts_home.dart';
 import '../screens/users/users_screen.dart';
 
-export 'package:auto_route/auto_route.dart';
 
 @MaterialAutoRouter(
   generateNavigationHelperExtension: true,
   routes: <AutoRoute>[
     CupertinoRoute(page: HomeScreen, initial: true),
     MaterialRoute(
-      path: '/users/:id?',
+      path: '/users/:id',
       page: UsersScreen,
       name: 'usersScreen',
       children: [
