+++ example/lib/screens/home_screen.dart
@@ -1,7 +1,7 @@
+import 'package:auto_route/auto_route.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 
-import '../router/router.dart';
 
 class HomeScreen extends StatelessWidget {
   const HomeScreen({Key key,String x}) :super(key: key);
@@ -18,6 +18,7 @@ class HomeScreen extends StatelessWidget {
               child: FlatButton(
                 child: Text("Users Screen"),
                 onPressed: () async {
+                  ExtendedNavigator.of(context);
 //                  context.navigator.router.findMatch(settings)
                   context.navigator.push("/users/23/posts?foo=bar");
                 },
