+++ example/lib/screens/users/sub/sub/posts_details.dart
@@ -1,6 +1,5 @@
 import 'package:flutter/material.dart';
-
-import '../../../../router/router.dart';
+import 'package:auto_route/auto_route.dart';
 
 class PostDetails extends StatelessWidget {
   @override
